+++ src/num/rand/mod.rs
@@ -31,14 +31,17 @@ crate::items! {
     mod xabc;
 }
 
-crate::items! { // structural access: _mods, _all
+crate::items! { // structural access: _mods, _internals, _all
     #[allow(unused)]
-    pub use _mods::*;
+    pub use {_mods::*, _internals::*};
 
     mod _mods {
         pub use super::xorshift::*;
         #[cfg(feature = "rand")]
         pub use super::{lgc::*, xabc::*, xoroshiro::*, xyza8::*};
+    }
+    pub(super) mod _internals { #![allow(unused)]
+        pub(crate) use super::xorshift::xorshift_basis;
     }
     pub(super) mod _all {
         #[doc(inline)]