+++ src/num/rand/xorshift/u128p.rs
@@ -30,8 +30,6 @@ impl ConstDefault for XorShift128p {
 impl XorShift128p {
     const DEFAULT_SEED: [u64; 2] = [0xDEFA_0017_DEFA_0017; 2];
 
-    #[cold] #[rustfmt::skip]
-    const fn cold_path_result() -> Option<Self> { None }
     #[cold] #[allow(dead_code)] #[rustfmt::skip]
     const fn cold_path_default() -> Self { Self::new_unchecked(Self::DEFAULT_SEED) }
 }
@@ -39,13 +37,12 @@ impl XorShift128p {
 impl XorShift128p {
     /// Returns a seeded `XorShift128+` generator from the given 2 × 64-bit seeds.
     ///
-    /// Returns `None` if all given seeds are `0`.
-    #[must_use]
-    pub const fn new(seeds: [u64; 2]) -> Option<Self> {
+    /// If the seed is `0`, the default seed is used instead.
+    pub const fn new(seeds: [u64; 2]) -> Self {
         if (seeds[0] | seeds[1]) == 0 {
-            Self::cold_path_result()
+            Self::cold_path_default()
         } else {
-            Some(Self(seeds))
+            Self(seeds)
         }
     }
 
@@ -107,14 +104,14 @@ impl XorShift128p {
     /// The seeds will be split in little endian order.
     #[cfg(feature = "split")]
     #[cfg_attr(feature = "nightly_doc", doc(cfg(feature = "split")))]
-    pub const fn new1_u128(seed: u128) -> Option<Self> {
+    pub const fn new1_u128(seed: u128) -> Self {
         Self::new(Cast(seed).into_u64_le())
     }
 
     /// Returns a seeded `XorShift128+` generator from the given 2 × 64-bit seeds.
     ///
     /// This is an alias of [`new`][Self#method.new].
-    pub const fn new2_u64(seeds: [u64; 2]) -> Option<Self> {
+    pub const fn new2_u64(seeds: [u64; 2]) -> Self {
         Self::new(seeds)
     }
 
@@ -123,7 +120,7 @@ impl XorShift128p {
     /// The seeds will be joined in little endian order.
     #[cfg(feature = "join")]
     #[cfg_attr(feature = "nightly_doc", doc(cfg(feature = "join")))]
-    pub const fn new4_u32(seeds: [u32; 4]) -> Option<Self> {
+    pub const fn new4_u32(seeds: [u32; 4]) -> Self {
         Self::new([
             Cast::<u64>::from_u32_le([seeds[0], seeds[1]]),
             Cast::<u64>::from_u32_le([seeds[2], seeds[3]]),
@@ -135,7 +132,7 @@ impl XorShift128p {
     /// The seeds will be joined in little endian order.
     #[cfg(feature = "join")]
     #[cfg_attr(feature = "nightly_doc", doc(cfg(feature = "join")))]
-    pub const fn new8_u16(seeds: [u16; 8]) -> Option<Self> {
+    pub const fn new8_u16(seeds: [u16; 8]) -> Self {
         Self::new([
             Cast::<u64>::from_u16_le([seeds[0], seeds[1], seeds[2], seeds[3]]),
             Cast::<u64>::from_u16_le([seeds[4], seeds[5], seeds[6], seeds[7]]),
@@ -145,7 +142,7 @@ impl XorShift128p {
     /// Returns a seeded `XorShift128` generator from the given 4 × 8-bit seeds.
     ///
     /// The seeds will be joined in little endian order.
-    pub const fn new16_u8(seeds: [u8; 16]) -> Option<Self> {
+    pub const fn new16_u8(seeds: [u8; 16]) -> Self {
         let s = seeds;
         Self::new([
             u64::from_le_bytes([s[0], s[1], s[2], s[3], s[4], s[5], s[6], s[7]]),